requirements.txt
Django==3.2.12
django-cleanup==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==9.0.1
requests==2.26.0
uwsgi
psycopg2==2.8.6
django-debug-toolbar==3.2.4
django-cors-headers
python-telegram-bot==13.11
django-admin-honeypot==1.1.0